requirements.txt
amqp==2.6.1
asgiref==3.2.10
attrs==20.1.0
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
Django==3.1
django-admin==2.0.1
django-celery-results==1.2.1
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-six==1.0.4
djangorestframework==3.11.1
gunicorn==20.0.4
idna==2.10
iniconfig==1.0.1
kombu==4.6.11
more-itertools==8.5.0
mysqlclient==2.0.1
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.0.1
pytz==2020.1
requests==2.24.0
screen==1.0.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
vine==1.3.0
xlwt==1.3.0
